Introduce InputPackets, that extracts messages from packets
