+++ src/protocol/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod message;
 pub(crate) mod input_message;
 pub(crate) mod slsk_buffer;
+pub(crate) mod packet;
